better windowed cross validationscript

+++ cross_validation_winsizes.sh
@@ -8,9 +8,18 @@ for i in ${plist[@]}; do
     echo "Working on window size $i"
     mlist=("top" "avg" "low")
     for m in ${mlist[@]}; do
-        python models/cross_validation.py $input_clusters_folder --out_dir $output_processing_folder --model forest_win_$i --methods $m --split_size $split --preprocess nominal_windowed --win_side_len $i
-        for ((j = 0; j <= split; j++)); do
-            rm "$output_processing_folder/fold$j/forest_win_${i}_${m}.params"
+        missing_file="F"
+        for ((j = 0; j <= split-1; j++)); do
+            if [ ! -f "$output_processing_folder/fold$j/forest_win_${i}_${m}.score" ]; then
+                echo "$output_processing_folder/fold$j/forest_win_${i}_${m}.score"
+                missing_file="T"
+            fi
         done
+        if [[ $missing_file == "T" ]]; then
+            python models/cross_validation.py $input_clusters_folder --out_dir $output_processing_folder --model forest_win_$i --methods $m --split_size $split --preprocess nominal_windowed --win_side_len $i
+            for ((j = 0; j <= split; j++)); do
+                rm "$output_processing_folder/fold$j/forest_win_${i}_${m}.params"
+            done
+        fi
     done
 done
